consolidate scene and quest editor
also moved node editors in their own folder
