+++ WolvenKit/Forms/frmCR2WDocument.cs
@@ -16,7 +16,6 @@ namespace WolvenKit
         public frmImagePreview ImageViewer;
         public Render.frmRender RenderViewer;
         private CR2WFile file;
-        public frmQuestFlowDiagram questflowDiagram;
 
 
         public DockPanel FormPanel => dockPanel;
