+++ WolvenKit/WolvenKit.csproj
@@ -315,21 +315,51 @@
     <Compile Include="AssemblyExtensions.cs" />
     <Compile Include="DDSUtility.cs" />
     <Compile Include="discord.cs" />
-    <Compile Include="FlowTreeEditors\quest\QuestPhaseEditor.cs" />
-    <Compile Include="FlowTreeEditors\quest\QuestLinkEditor.cs" />
-    <Compile Include="FlowTreeEditors\quest\QuestPhaseInputEditor.cs" />
-    <Compile Include="FlowTreeEditors\quest\QuestScriptBlockEditor.cs" />
+    <Compile Include="FlowTreeEditors\LinkEditor.cs" />
+    <Compile Include="FlowTreeEditors\Quest\QuestPhaseEditor.cs" />
+    <Compile Include="FlowTreeEditors\Quest\QuestPhaseInputEditor.cs" />
+    <Compile Include="FlowTreeEditors\Quest\QuestScriptBlockEditor.cs" />
+    <Compile Include="FlowTreeEditors\Scene\SceneChoiceEditor.cs">
+      <SubType>UserControl</SubType>
+    </Compile>
+    <Compile Include="FlowTreeEditors\Scene\SceneChoiceEditor.Designer.cs">
+      <DependentUpon>SceneChoiceEditor.cs</DependentUpon>
+    </Compile>
+    <Compile Include="FlowTreeEditors\Scene\SceneFlowConditionEditor.cs">
+      <SubType>UserControl</SubType>
+    </Compile>
+    <Compile Include="FlowTreeEditors\Scene\SceneFlowConditionEditor.Designer.cs">
+      <DependentUpon>SceneFlowConditionEditor.cs</DependentUpon>
+    </Compile>
+    <Compile Include="FlowTreeEditors\Scene\SceneLinkEditor.cs">
+      <SubType>UserControl</SubType>
+    </Compile>
+    <Compile Include="FlowTreeEditors\Scene\SceneLinkEditor.Designer.cs">
+      <DependentUpon>SceneLinkEditor.cs</DependentUpon>
+    </Compile>
+    <Compile Include="FlowTreeEditors\Scene\SceneRandomizerEditor.cs">
+      <SubType>UserControl</SubType>
+    </Compile>
+    <Compile Include="FlowTreeEditors\Scene\SceneRandomizerEditor.Designer.cs">
+      <DependentUpon>SceneRandomizerEditor.cs</DependentUpon>
+    </Compile>
+    <Compile Include="FlowTreeEditors\Scene\SceneSectionEditor.cs">
+      <SubType>UserControl</SubType>
+    </Compile>
+    <Compile Include="FlowTreeEditors\Scene\SceneSectionEditor.Designer.cs">
+      <DependentUpon>SceneSectionEditor.cs</DependentUpon>
+    </Compile>
     <Compile Include="Forms\frmLoading.cs">
       <SubType>Form</SubType>
     </Compile>
     <Compile Include="Forms\frmLoading.Designer.cs">
       <DependentUpon>frmLoading.cs</DependentUpon>
     </Compile>
-    <Compile Include="Forms\frmQuestFlowDiagram.cs">
+    <Compile Include="Forms\frmChunkFlowDiagram.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="Forms\frmQuestFlowDiagram.Designer.cs">
-      <DependentUpon>frmQuestFlowDiagram.cs</DependentUpon>
+    <Compile Include="Forms\frmChunkFlowDiagram.Designer.cs">
+      <DependentUpon>frmChunkFlowDiagram.cs</DependentUpon>
     </Compile>
     <Compile Include="Forms\frmStringsGuiImporterIDDialog.cs">
       <SubType>Form</SubType>
@@ -446,36 +476,6 @@
       <DependentUpon>ChunkEditor.cs</DependentUpon>
     </Compile>
     <Compile Include="FlowTreeEditors\MoveEditorArgs.cs" />
-    <Compile Include="FlowTreeEditors\SceneChoiceEditor.cs">
-      <SubType>UserControl</SubType>
-    </Compile>
-    <Compile Include="FlowTreeEditors\SceneChoiceEditor.Designer.cs">
-      <DependentUpon>SceneChoiceEditor.cs</DependentUpon>
-    </Compile>
-    <Compile Include="FlowTreeEditors\SceneFlowConditionEditor.cs">
-      <SubType>UserControl</SubType>
-    </Compile>
-    <Compile Include="FlowTreeEditors\SceneFlowConditionEditor.Designer.cs">
-      <DependentUpon>SceneFlowConditionEditor.cs</DependentUpon>
-    </Compile>
-    <Compile Include="FlowTreeEditors\SceneLinkEditor.cs">
-      <SubType>UserControl</SubType>
-    </Compile>
-    <Compile Include="FlowTreeEditors\SceneLinkEditor.Designer.cs">
-      <DependentUpon>SceneLinkEditor.cs</DependentUpon>
-    </Compile>
-    <Compile Include="FlowTreeEditors\SceneRandomizerEditor.cs">
-      <SubType>UserControl</SubType>
-    </Compile>
-    <Compile Include="FlowTreeEditors\SceneRandomizerEditor.Designer.cs">
-      <DependentUpon>SceneRandomizerEditor.cs</DependentUpon>
-    </Compile>
-    <Compile Include="FlowTreeEditors\SceneSectionEditor.cs">
-      <SubType>UserControl</SubType>
-    </Compile>
-    <Compile Include="FlowTreeEditors\SceneSectionEditor.Designer.cs">
-      <DependentUpon>SceneSectionEditor.cs</DependentUpon>
-    </Compile>
     <Compile Include="Forms\frmAddChunk.cs">
       <SubType>Form</SubType>
     </Compile>
@@ -493,12 +493,6 @@
     </Compile>
     <Compile Include="Forms\frmAssetBrowser.Designer.cs">
       <DependentUpon>frmAssetBrowser.cs</DependentUpon>
-    </Compile>
-    <Compile Include="Forms\frmChunkFlowDiagram.cs">
-      <SubType>Form</SubType>
-    </Compile>
-    <Compile Include="Forms\frmChunkFlowDiagram.Designer.cs">
-      <DependentUpon>frmChunkFlowDiagram.cs</DependentUpon>
     </Compile>
     <Compile Include="Forms\frmChunkList.cs">
       <SubType>Form</SubType>
@@ -581,18 +575,18 @@
     <Compile Include="SortableBindinglist.cs" />
     <Compile Include="Win32.cs" />
     <Compile Include="WKPackage.cs" />
+    <EmbeddedResource Include="FlowTreeEditors\Scene\SceneFlowConditionEditor.resx">
+      <DependentUpon>SceneFlowConditionEditor.cs</DependentUpon>
+    </EmbeddedResource>
+    <EmbeddedResource Include="FlowTreeEditors\Scene\SceneLinkEditor.resx">
+      <DependentUpon>SceneLinkEditor.cs</DependentUpon>
+    </EmbeddedResource>
     <EmbeddedResource Include="Forms\frmAbout.resx">
       <DependentUpon>frmAbout.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="FlowTreeEditors\ChunkEditor.resx">
       <DependentUpon>ChunkEditor.cs</DependentUpon>
       <SubType>Designer</SubType>
-    </EmbeddedResource>
-    <EmbeddedResource Include="FlowTreeEditors\SceneFlowConditionEditor.resx">
-      <DependentUpon>SceneFlowConditionEditor.cs</DependentUpon>
-    </EmbeddedResource>
-    <EmbeddedResource Include="FlowTreeEditors\SceneLinkEditor.resx">
-      <DependentUpon>SceneLinkEditor.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Forms\frmAddChunk.resx">
       <DependentUpon>frmAddChunk.cs</DependentUpon>
@@ -605,9 +599,6 @@
     <EmbeddedResource Include="Forms\frmAssetBrowser.resx">
       <DependentUpon>frmAssetBrowser.cs</DependentUpon>
       <SubType>Designer</SubType>
-    </EmbeddedResource>
-    <EmbeddedResource Include="Forms\frmChunkFlowDiagram.resx">
-      <DependentUpon>frmChunkFlowDiagram.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Forms\frmChunkList.resx">
       <DependentUpon>frmChunkList.cs</DependentUpon>
@@ -660,8 +651,8 @@
     <EmbeddedResource Include="Forms\frmPackSettings.resx">
       <DependentUpon>frmPackSettings.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="Forms\frmQuestFlowDiagram.resx">
-      <DependentUpon>frmQuestFlowDiagram.cs</DependentUpon>
+    <EmbeddedResource Include="Forms\frmChunkFlowDiagram.resx">
+      <DependentUpon>frmChunkFlowDiagram.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Forms\frmRenameDialog.resx">
       <DependentUpon>frmRenameDialog.cs</DependentUpon>
